Lots more Buzz options.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/0.05@1900 0a995953-517f-4c62-b348-104708d926b2

+++ GUI/spot-on-a.cc
@@ -3075,9 +3075,7 @@ void spoton::slotSetPassphrase(void)
   QByteArray derivedKey
     (spoton_crypt::derivedKey(m_ui.cipherType->currentText(),
 			      m_ui.hashType->currentText(),
-			      static_cast<unsigned long> (m_ui.
-							  iterationCount->
-							  value()),
+			      m_ui.iterationCount->value(),
 			      str1,
 			      salt,
 			      error1));
@@ -3357,9 +3355,7 @@ void spoton::slotValidatePassphrase(void)
 	QByteArray key
 	  (spoton_crypt::derivedKey(m_ui.cipherType->currentText(),
 				    m_ui.hashType->currentText(),
-				    static_cast
-				    <unsigned long> (m_ui.
-						     iterationCount->value()),
+				    m_ui.iterationCount->value(),
 				    m_ui.passphrase->text(),
 				    salt,
 				    error));
